% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iex.R
\name{iex_api}
\alias{iex_api}
\title{Base method that sends an HTTP GET request to the REST endpoint.  This returns
an S3 object of class \code{iex_api} which has three accessible fields:
\code{path}, \code{response} and \code{content} containing the API path, the unparsed API
response and the parsed content from the API's response (the latter usually
being a list). Note that this package causes R to pause 0.2 seconds after
executing an API call to avoid the user being throttled by the API (which
enforces a 5 request per second limit)}
\usage{
iex_api(path, query = NULL, version = "1.0")
}
\description{
Base method that sends an HTTP GET request to the REST endpoint.  This returns
an S3 object of class \code{iex_api} which has three accessible fields:
\code{path}, \code{response} and \code{content} containing the API path, the unparsed API
response and the parsed content from the API's response (the latter usually
being a list). Note that this package causes R to pause 0.2 seconds after
executing an API call to avoid the user being throttled by the API (which
enforces a 5 request per second limit)
}
