% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iex.R
\name{iex_chart}
\alias{iex_chart}
\title{Returns historical chart data for the stock}
\usage{
iex_chart(symbol, timeframe = "5y", query = NULL, simplify = TRUE)
}
\description{
Returns historical chart data for the stock
}
