% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iex.R
\name{iex_dividends}
\alias{iex_dividends}
\title{Returns all dividends declared by a company over a given timeframe. The
default time frame is 5 years.}
\usage{
iex_dividends(symbol, timeframe = "5y", simplify = TRUE)
}
\description{
Returns all dividends declared by a company over a given timeframe. The
default time frame is 5 years.
}
