% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iex.R
\name{iex_earnings}
\alias{iex_earnings}
\title{Returns earnings data from the four most recent reported quarters.}
\usage{
iex_earnings(symbol, simplify = TRUE)
}
\description{
Returns earnings data from the four most recent reported quarters.
}
