% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iex.R
\name{iex_financials}
\alias{iex_financials}
\title{Pulls income statement, balance sheet, and cash flow data from the four most
recent reported quarters or years}
\usage{
iex_financials(symbol, period = c("quarter", "annual"), simplify = TRUE)
}
\description{
Pulls income statement, balance sheet, and cash flow data from the four most
recent reported quarters or years
}
