% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iex.R
\name{iex_listed_symbols}
\alias{iex_listed_symbols}
\title{Returns information about companies listed on the IEX exchange.
NOTE: Only 4 companies are returned so far (and 2 of these appear to be
tests).}
\usage{
iex_listed_symbols(simplify = TRUE)
}
\description{
Returns information about companies listed on the IEX exchange.
NOTE: Only 4 companies are returned so far (and 2 of these appear to be
tests).
}
