% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iex.R
\name{iex_market_previous}
\alias{iex_market_previous}
\title{This returns previous day adjusted price data for all stocks traded on IEX.}
\usage{
iex_market_previous(simplify = TRUE)
}
\description{
This returns previous day adjusted price data for all stocks traded on IEX.
}
