% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iex.R
\name{iex_sector_performance}
\alias{iex_sector_performance}
\title{This returns an array of each sector and performance for the current trading
day. Performance is based on each sector ETF.}
\usage{
iex_sector_performance(simplify = TRUE)
}
\description{
This returns an array of each sector and performance for the current trading
day. Performance is based on each sector ETF.
}
