% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iex.R
\name{iex_splits}
\alias{iex_splits}
\title{Returns all splits declared by a company over a given timeframe. The
default time frame is 5 years.}
\usage{
iex_splits(symbol, timeframe = "5y", simplify = TRUE)
}
\description{
Returns all splits declared by a company over a given timeframe. The
default time frame is 5 years.
}
