% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iex.R
\name{iex_stock_peers}
\alias{iex_stock_peers}
\title{An array of peer tickers as defined by IEX. This is not intended to
represent a definitive or accurate list of peers, and is subject to
change at any time.}
\usage{
iex_stock_peers(symbol, simplify = TRUE)
}
\description{
An array of peer tickers as defined by IEX. This is not intended to
represent a definitive or accurate list of peers, and is subject to
change at any time.
}
