% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iex.R
\name{iex_stock_previous}
\alias{iex_stock_previous}
\title{This returns previous day adjusted price data for a single stock.}
\usage{
iex_stock_previous(symbol, simplify = TRUE)
}
\description{
This returns previous day adjusted price data for a single stock.
}
