% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iex.R
\name{iex_stock_relevant}
\alias{iex_stock_relevant}
\title{Similar to the peers endpoint, except this will return most active market
symbols when peers are not available.}
\usage{
iex_stock_relevant(symbol, simplify = TRUE)
}
\description{
Similar to the peers endpoint, except this will return most active market
symbols when peers are not available.
}
