% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iex.R
\name{iex_traded_symbols}
\alias{iex_traded_symbols}
\title{Returns a list of all stock symbols traded on the IEX exchange.}
\usage{
iex_traded_symbols(simplify = TRUE)
}
\arguments{
\item{simplify}{Set to FALSE to get the raw API response.  When TRUE
(the default) we simplify the result to a data.frame}
}
\value{
A data frame containing the following columns:
symbol - Stock symbol
name - Stock name
date - Most recent date (?)
isEnabled
type - "cs" = "common stock"
iexId - Unique ID for the stock on the IEX exchange
}
\description{
Returns a list of all stock symbols traded on the IEX exchange.
}
