% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iex.R
\name{list_to_df}
\alias{list_to_df}
\title{Converts a list of lists to a data.frame. This helps parse the typical
response$content list returned by most IEX functions.}
\usage{
list_to_df(content, ensure_columns = NULL)
}
\description{
Converts a list of lists to a data.frame. This helps parse the typical
response$content list returned by most IEX functions.
}
