% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iex.R
\name{market}
\alias{market}
\title{Market volume data}
\usage{
market(fields = NULL, version = "1.0")
}
\arguments{
\item{fields}{a vector of fields names to return (case sensitive). When
set to \code{NULL} (default) returns values for all fields.}

\item{version}{the API version number (default: \code{"1.0"}) which is used to
define the API URL.}
}
\value{
an S3 object of class \code{iex_api} which has three accessible fields:
\code{path}, \code{response} and \code{content} containing the API path, the unparsed API
response and the parsed content from the API's response (the latter usually
being a list). Note that this package causes R to pause 0.2 seconds after
executing an API call to avoid the user being throttled by the API (which
enforces a 5 request per second limit)
}
\description{
Provides exchange trade volume data in near real time.
}
\examples{
\dontrun{
market()
}
}
\references{
\href{https://iextrading.com/developer/#market-market}{IEX market API documentation}
}
