% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iex.R
\name{stats}
\alias{stats}
\alias{intraday_stats}
\alias{recent_stats}
\alias{records_stats}
\alias{monthly_stats}
\alias{daily_stats}
\title{Trading stats}
\usage{
intraday_stats(fields = NULL, version = "1.0")

recent_stats(fields = NULL, version = "1.0")

records_stats(fields = NULL, version = "1.0")

monthly_stats(date = NULL, fields = NULL, version = "1.0")

daily_stats(date = NULL, last = NULL, fields = NULL, version = "1.0")
}
\arguments{
\item{fields}{a vector of fields names to return (case sensitive). When
set to \code{NULL} (default) returns values for all fields.}

\item{version}{the API version number (default: \code{"1.0"}) which is used to
define the API URL.}

\item{date}{should be a string of the format \code{"YYYYMM"} or \code{"YYYYMMDD"} which
is a valid option for \code{daily_stats()} or \code{NULL} (default) which returns the
prior trading date's data for \code{daily_stats()}  and the prior month's
trading data for \code{monthly_stats()}}

\item{last}{can be used in place of \code{date} to retrieve the last \code{n} number of
trading days' data. If this is supplied, any value supplied to \code{date} is
ignored.}
}
\value{
an S3 object of class \code{iex_api} which has three accessible fields:
\code{path}, \code{response} and \code{content} containing the API path, the unparsed API
response and the parsed content from the API's response (the latter usually
being a list). Note that this package causes R to pause 0.2 seconds after
executing an API call to avoid the user being throttled by the API (which
enforces a 5 request per second limit)
}
\description{
A set of functions that return trading statistics.
}
\references{
\href{https://iextrading.com/developer/#stats}{IEX stats API documentation}
}
